requirements.txt
PyYAML==5.3.1
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
seaborn==0.11.2
six==1.14.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cu113
torchvision==0.11.1+cu113